Move autopost specification to integration folder. Move expected result representations to separate files.
